Move/remove model cards and other minor tidy ups (#422)
* Remove sensor. Tidy. Refactor model card upload
